Modified DialogFragment so that floating point numbers can be entered, while ProdTransactionData object contains float primitives for numbers entered in the DialogFragment the DialogFragment will accept double precision floating numbers as well it seems exceeding the limit of 2 to the power of 32 is not a problem.

+++ app/src/main/java/android/apex/ge/coffee/Fragments/PageRecViewFragment.java
@@ -253,7 +253,7 @@ public class PageRecViewFragment extends Fragment implements ILibObjectCrud<Prod
 
     @Override
     public void onFinishEditDialog(String inputNumber1, String inputNumber2, String prodPPID) {
-        Log.d(LOG_TAG, "Numbers entered were: \n" + Integer.parseInt(inputNumber1) + "\n" + Integer.parseInt(inputNumber2));
+        Log.d(LOG_TAG, "Numbers entered were: \n" + Double.parseDouble(inputNumber1) + "\n" + Double.parseDouble(inputNumber2));
 
         createAndAddProdTransactionData(inputNumber1, inputNumber2, prodPPID);
